Switch to kubeadm
Before this point, the CIs were using the local-up-cluster.sh script
to start the Kubernetes cluster (local-up-cluster.sh is a script provided
by the kubernetes project, that run the kubelet and kube-api together. It
is mainly used for testing). This is bad since in production, kubeadm is
usually used, also kubeadm is easier to run and maintain, and provide the
ability to do multi-node CIs in the future.