Collect logs of pods and services
Before this point logs were collected in a Jenkins build step that
would look for logs inside the /tmp dir. This is where the local_cluster_up.sh
used to save the logs. Since moving to kubeadm, the logs are now split
between the kube-api pod, and the kubelet service. This patch collects
the logs from all the pods deployed in the system, and from the kubelet
and docker service and adds them to the LOGDIR directory.