Add missing CNI_BIN_DIR variable
Added the missing CNI_BIN_DIR and CNI_CONF_DIR variable to the
common_functions script, it seems it was mistakenly deleted in #52.
The CNI_BIN_DIR is used to copy the networking container plugins
binaries to the correct place. This caused an issue that the pods
were not finding the macvlan bin.