Disable deleting docker images every job
Because of DockerHub new pull rate limit on anonymous users, the
CIs started to fail because it was hitting the limit when trying
to pull the images for the Kubernetes pods. This patch disables
the docker images cleaning, so that the CIs will minimize the number
of pulls it will perform when triggered.