Unload the mlx_compat after deleting the nic-cluster-policy
Unload the mlx_compat module after deleteng the nic-cluster-policy,
this is needed to be able to load the inbox rdma modules after running
the MOFED container. It is done manually because of an issue in the network
operator project where in some systems, the mlx_compat module is not
removed as part of the MOFED container cleanup.