Unlabel the master label of the node before deploying helm
Following Mellanox/network-operator#159, the deployment order for
the nic-cluster-policy changed, and the secondary network components
are now deployed first. Because of that, the node must not have the
`node-role.kubernetes.io/master` label to it for the multus to be
deployed successfully. Since the scripts were waiting for the OFED
state to become ready for it to unlabel the node, the helm check
started to fail, but because there is a bug in the scripts that
cause the scripts to not fail on this step, it went unnoticed.

This patch unlable the node after the helm deploy command for it
to deploy the multus successfully and not failing the deploy step.