Use nic-cluster-policy to configure multus in nic_operator CI
Before this point, multus was manually deployed in the nic-operator
project, and only secondary network test used to use the secondary
network field of the nic-cluster-policy. This was the case since
the CI was scripted and before the secondary network feature was
added to the project.

As part of CI enhacment and to test all the operator component using
the seconady network, this patch deploys the seconadry network components
on all tests and remove the normal multus deployment from the nic-operator
scripts.