Prebuild the MOFED container to match host kernel
This patch modifies the MOFED container image so that it would have
the DKMS already done. This will speed up the setup time of the
MOFED container significantly. This is needed for the kind deployment
of the network operator, since it takes longer times for the kind worker
containers to build the driver which may fail in some cases.