Add support for using a per project environment file for ansible projects
Added the ability to configure a per project environment.sh file
that would be sourced at the start of the run_kind_ci.sh. This is
done to be able to configure the default values of the run_kind_ci.sh
script per project. This is needed since some projects wont work as
expected with the run_kind_ci.sh default values, and changing that
can break other CIs.