Restart the SRIOV operator config daemon before testing
Sometimes, SRIOV operator E2E test fails because it fails to configure
VFs, debuging some more, it was observed that when that happens, the
SRIOV operator config daemon fails to discover the PFs of the machine.

This patch workaround that issue by restarting the SRIOV operator
config daemon so that it will rediscover the interfaces.