Support deploying the GPU operator for kind CIs
Added the ability to deploy the GPU operator for kind node. To do
that, kind node OS version must be taken into considaration when
pulling the image, and since an image based on ubuntu20.10 is not
currently available for both the nv-peer-mem driver image, and
the GPU driver image, a workaround to retag the images from ubuntu20.04
to ubuntu20.10 and upload them to kind nodes is currently adopted.