Get interface RDMA resource for ib_write_bw command
RDMA resources are being fetched for the ib_write_bw based on alphabitic
order of the RDMA devices inside the testing pod. This in normal cases,
is not an issue, but in case of mixed ethernet and infiniband setup, this
interduces an issue that the ib_write_bw may take the RDMA resource for
an infiniband interface when it is not requested or the interface is down,
because the opensm service is down.

This patch fix that by fetching the RDMA resource name of the interface
that was configured in the rdma-shared-device-plugin.