Added the Copy rights validator
Added a CI to validate copyrights of added files of a specific git
project. It checks if the PR's newly added files have the right copy
rights header (`$(date +%Y) NVIDIA CORPORATION & AFFILIATES`).

Currently the CI looks for the header in all the project's newly added
files regardless of the file type.