Workaround setting the vfs macs for the sriov-operator sriovs VFs
Sometimes sriov-network-operator inside network-operator in kind
cluster fails to set the VFs node GUIDS, this will cause the RDMA
traffic to fail between pods that use these VFs.

This patch workaround this by setting the VFs node GUIDs manually.
This workaround should not be nessacary in master sriov-network-operator
images, so it should be removed when a new image of the sriov-operator
is applied in the network-operator CIs,