Fix pending IB interfaces for host device IB test
nic-operator-kind CI has a race condition between the vf-switcher and
the sriov-device-plugin for discovering IB interfaces. Basically the sriov
device plugin would finish its work before the vf-switcher switches all IB
interfaces into the KIND container. Two scenarios happen here, one is with
MOFED, and the other is without MOFED. If the nic--cluster-policy is with
MOFED, then there is no way of solving the race, since the mofed do a driver
reload, then immediately deploys the sriov-device-plugin.

This patch fix non MOFED case by adding a delay after loading the ib_ipoib module
to give the interface some time to load, and also move the MOFED test to the
nic-operator project (which does not have kind), to not face issues with the
vf-switcher service race.