Fix missing sudo for ovn-kubernetes build command
Following ovn patch [1] being merged, the way the Makefile use
the OCI to build the images changed, which broke the CI since it looks
for `docker build` and change it to `sudo docker build`. This patch fix
that by utilizing the patch new ENV variable `OCI_BIN` to set the command
to sudo docker.

[1] https://github.com/ovn-org/ovn-kubernetes/pull/2315