Support changing the upstream PR ref for copyright CI
Supported changing the upstream PR ref name to support running the
CI for different GIT services, and not only github.

Also specified the PR merge reference because the clone would fail
with globs in case of gitlab, and also to only test the new change
files instead of checking the branch. This would remove the need to
rebase the source branch if it was an old branch that would have already
deleted files in the master.