Restart sriov-network-device-plugin after creating sriov node policy
Using latest sriov operator images with the nic-operator-helm project
breaks the CI, it was observed that the sriov-network-device-plugin is
not restarted after successfully creating the policy causing it to not
register the newly created VFs.

This patch workaround the issue by manually restarting the sriov-network
device plugin after creating the sriovnetworknodepolicy.